requirements.txt
chardet==4.0.0
click==8.0.0
colorama==0.4.4
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==3.0.0
MarkupSafe==2.0.0rc2
mysql-connector-python==8.0.24
mysqlclient==2.0.3
peewee==3.14.4
pinyin==0.4.0
protobuf==3.16.0
pypinyin==0.41.0
python-dateutil==2.8.1
python-frontmatter==1.0.0
PyYAML==5.1
six==1.16.0
Werkzeug==2.0.0
